+++ src/application/Game.java
@@ -73,7 +73,7 @@ public class Game {
 		// entities.get(0).sndSpawn.startSound();
 	}
 
-	public void MainThreadFunctions(long time, long dtime) {
+	public void MainThreadFunctions(double time, double dtime) {
 		Main.game.renderer.render(time, dtime);
 		// SoundHandler.play(Soundlist);
 	}
